[onert] Move exec and interp unittest
This commit moves exec and interp unittest into "src/".
