+++ runtime/onert/core/src/exec/feature/nchw/View.test.cc
@@ -14,15 +14,15 @@
  * limitations under the License.
  */
 
-#include "../MockTensor.h"
+#include "View.h"
 
-#include "exec/feature/nhwc/Reader.h"
+#include "../MockTensor.h"
 
 #include <gtest/gtest.h>
 
 using namespace onert::exec::feature;
 
-template <typename T> class Reader_nhwc : public testing::Test
+template <typename T> class View_nchw : public testing::Test
 {
 public:
   void setData(std::initializer_list<T> list) { _data = std::make_shared<std::vector<T>>(list); }
@@ -39,20 +39,20 @@ public:
                                       width * elem_size);
   }
 
-  void createReader()
+  void createView()
   {
-    _reader =
-      std::make_shared<nhwc::Reader<T>>(_shape, _stride, _data->data(), _data->size() * sizeof(T));
+    _view =
+      std::make_shared<nchw::View<T>>(_shape, _stride, _data->data(), _data->size() * sizeof(T));
   }
 
   void createUsingMockTensor()
   {
     onert::ir::Shape shape = {_shape.N, _shape.H, _shape.W, _shape.C};
-    _tensor = std::make_shared<MockTensor<T>>(shape, _data->data(), onert::ir::Layout::NHWC);
-    _reader = std::make_shared<nhwc::Reader<T>>(_tensor.get());
+    _tensor = std::make_shared<MockTensor<T>>(shape, _data->data(), onert::ir::Layout::NCHW);
+    _view = std::make_shared<nchw::View<T>>(_tensor.get());
   }
 
-  std::shared_ptr<nhwc::Reader<T>> _reader = nullptr;
+  std::shared_ptr<nchw::View<T>> _view = nullptr;
 
 private:
   std::shared_ptr<std::vector<T>> _data = nullptr;
@@ -61,26 +61,25 @@ private:
   std::shared_ptr<MockTensor<T>> _tensor = nullptr;
 };
 
-using ReaderTypes = ::testing::Types<float, int32_t, uint8_t, int8_t, int16_t>;
-TYPED_TEST_SUITE(Reader_nhwc, ReaderTypes);
-TYPED_TEST_SUITE(MockTensorReader_nhwc, ReaderTypes);
+using ViewTypes = ::testing::Types<float, int32_t, uint8_t, int8_t, int16_t>;
+TYPED_TEST_SUITE(View_nchw, ViewTypes);
 
-TYPED_TEST(Reader_nhwc, basic_reader)
+TYPED_TEST(View_nchw, basic_view)
 {
   this->setData({0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11});
   this->setShape(1, 2, 3, 2);
-  this->setStride(12, 1, 6, 2);
-  this->createReader();
+  this->setStride(12, 6, 2, 1);
+  this->createView();
 
   // Data: NCHW
-  // Shape: NHWC
-  ASSERT_EQ(this->_reader->at(0, 1, 1, 0), 8);
-  ASSERT_EQ(this->_reader->at(1, 1, 0), 8);
+  // Shape: NCHW
+  ASSERT_EQ(this->_view->at(0, 1, 1, 0), 8);
+  ASSERT_EQ(this->_view->at(1, 1, 0), 8);
 
-  // Data: NHWC
-  // Shape: NHWC
+  // Data: NCHW
+  // Shape: NCHW
   this->createUsingMockTensor();
 
-  ASSERT_EQ(this->_reader->at(0, 1, 1, 0), 6);
-  ASSERT_EQ(this->_reader->at(1, 1, 0), 6);
+  ASSERT_EQ(this->_view->at(0, 1, 1, 0), 6);
+  ASSERT_EQ(this->_view->at(1, 1, 0), 6);
 }
