+++ runtime/onert/core/src/exec/feature/nhwc/Reader.test.cc
@@ -14,15 +14,15 @@
  * limitations under the License.
  */
 
-#include "../MockTensor.h"
+#include "Reader.h"
 
-#include "exec/feature/nhwc/View.h"
+#include "../MockTensor.h"
 
 #include <gtest/gtest.h>
 
 using namespace onert::exec::feature;
 
-template <typename T> class View_nhwc : public testing::Test
+template <typename T> class Reader_nhwc : public testing::Test
 {
 public:
   void setData(std::initializer_list<T> list) { _data = std::make_shared<std::vector<T>>(list); }
@@ -39,20 +39,20 @@ public:
                                       width * elem_size);
   }
 
-  void createView()
+  void createReader()
   {
-    _view =
-      std::make_shared<nhwc::View<T>>(_shape, _stride, _data->data(), _data->size() * sizeof(T));
+    _reader =
+      std::make_shared<nhwc::Reader<T>>(_shape, _stride, _data->data(), _data->size() * sizeof(T));
   }
 
   void createUsingMockTensor()
   {
     onert::ir::Shape shape = {_shape.N, _shape.H, _shape.W, _shape.C};
     _tensor = std::make_shared<MockTensor<T>>(shape, _data->data(), onert::ir::Layout::NHWC);
-    _view = std::make_shared<nhwc::View<T>>(_tensor.get());
+    _reader = std::make_shared<nhwc::Reader<T>>(_tensor.get());
   }
 
-  std::shared_ptr<nhwc::View<T>> _view = nullptr;
+  std::shared_ptr<nhwc::Reader<T>> _reader = nullptr;
 
 private:
   std::shared_ptr<std::vector<T>> _data = nullptr;
@@ -61,26 +61,26 @@ private:
   std::shared_ptr<MockTensor<T>> _tensor = nullptr;
 };
 
-using ViewTypes = ::testing::Types<float, int32_t, uint8_t, int8_t, int16_t>;
-TYPED_TEST_SUITE(View_nhwc, ViewTypes);
-TYPED_TEST_SUITE(MockTensorView_nhwc, ViewTypes);
+using ReaderTypes = ::testing::Types<float, int32_t, uint8_t, int8_t, int16_t>;
+TYPED_TEST_SUITE(Reader_nhwc, ReaderTypes);
+TYPED_TEST_SUITE(MockTensorReader_nhwc, ReaderTypes);
 
-TYPED_TEST(View_nhwc, basic_view)
+TYPED_TEST(Reader_nhwc, basic_reader)
 {
   this->setData({0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11});
   this->setShape(1, 2, 3, 2);
   this->setStride(12, 1, 6, 2);
-  this->createView();
+  this->createReader();
 
   // Data: NCHW
   // Shape: NHWC
-  ASSERT_EQ(this->_view->at(0, 1, 1, 0), 8);
-  ASSERT_EQ(this->_view->at(1, 1, 0), 8);
+  ASSERT_EQ(this->_reader->at(0, 1, 1, 0), 8);
+  ASSERT_EQ(this->_reader->at(1, 1, 0), 8);
 
   // Data: NHWC
   // Shape: NHWC
   this->createUsingMockTensor();
 
-  ASSERT_EQ(this->_view->at(0, 1, 1, 0), 6);
-  ASSERT_EQ(this->_view->at(1, 1, 0), 6);
+  ASSERT_EQ(this->_reader->at(0, 1, 1, 0), 6);
+  ASSERT_EQ(this->_reader->at(1, 1, 0), 6);
 }
